refactor: isolate timer time section to prevent sidebar re-renders
